requirements.txt
openai-whisper
sounddevice
numpy
pyperclip
pynput
